Code refactoring, extract CameraModel class, create main.py script
