+++ main.py
@@ -1,0 +1,5 @@
+from CameraApplication import  CameraApplication
+
+CHESSBOARD_SIZE = (9, 6)
+app = CameraApplication()
+app.run()
