+++ parameters.txt
@@ -1,16 +1,32 @@
 RET
-0.13337499423750498
+0.13820758590135082
 MTX
-[[429.72890581   0.         315.89452422]
- [  0.         425.94741996 199.82777522]
+[[413.60870129   0.         316.14943298]
+ [  0.         408.16680394 200.53351205]
  [  0.           0.           1.        ]]
 DIST
-[[-0.28352422  0.17288007 -0.00817197  0.00195675 -0.22050108]]
+[[-0.27134371  0.23207296 -0.00716998  0.00187722 -0.44250936]]
 RVECS
-(array([[-0.04413168],
-       [ 0.48339346],
-       [ 2.98905721]]),)
+(array([[-0.04311639],
+       [ 0.46145548],
+       [ 2.99233777]]), array([[-0.04318103],
+       [ 0.46147372],
+       [ 2.99234108]]), array([[-0.0430899 ],
+       [ 0.46170562],
+       [ 2.99230812]]), array([[-0.04298019],
+       [ 0.46150006],
+       [ 2.99233804]]), array([[-0.04295998],
+       [ 0.46159246],
+       [ 2.99234694]]))
 TVECS
-(array([[ 5.30063235],
-       [ 2.98503483],
-       [13.381472  ]]),)
+(array([[ 5.29315079],
+       [ 2.97140222],
+       [12.88634647]]), array([[ 5.29318227],
+       [ 2.97144841],
+       [12.88647253]]), array([[ 5.29306658],
+       [ 2.97129637],
+       [12.88564058]]), array([[ 5.29309876],
+       [ 2.97132022],
+       [12.88576743]]), array([[ 5.29301201],
+       [ 2.9714003 ],
+       [12.88568655]]))
